Make use of bootstrap component for tabs and refactor code for the TabbedLayout component
